requirements.txt
fastapi
telethon
python-dotenv
uvicorn